refactor(handlers): merge getOrCreateThread twins into shared helper (#145)
getOrCreateThread (pr.ts) and getOrCreateIssueThread (issue.ts) were
line-for-line twins differing only in thread-name kind, DB persistence
method, and seed message text. Extract the shared recovery/creation
logic into getOrCreateMessageThread in src/discord/threads.ts,
parametrized by thread name, seed message, and a persistThreadId
callback. Both public wrappers keep their existing signatures and
export locations.

fetchAndUnarchiveThread moves to the new module; pr.ts re-exports it
for compatibility with existing importers.
