+++ src/handlers/issue.ts
@@ -7,9 +7,9 @@ import { StateDb, IssueMessage } from '../db/state.js';
 import { buildIssueEmbed, buildIssueClosedReply, buildIssueReopenedReply, buildThreadName, IssueData } from '../embeds/builders.js';
 import { getChannelForEvent, ChannelConfig } from '../config/channels.js';
 import { getExistingIssueMessage } from '../discord/lookup.js';
+import { getOrCreateMessageThread } from '../discord/threads.js';
 import { withRetry } from '../utils/retry.js';
-import { isThreadAlreadyCreatedError, isUnknownMessageError } from '../utils/discord-errors.js';
-import { fetchAndUnarchiveThread } from './pr.js';
+import { isUnknownMessageError } from '../utils/discord-errors.js';
 
 export interface IssueEventPayload {
   action: 'opened' | 'closed' | 'reopened' | 'labeled' | 'unlabeled' | 'edited';
@@ -166,40 +166,11 @@ export async function getOrCreateIssueThread(
   issue: IssueData,
   existing: IssueMessage
 ): Promise<ThreadChannel> {
-  // Thread ID == parent message ID; archived threads are invisible to the
-  // cache-only Message#thread, so try a direct fetch even without a threadId.
-  const threadId = existing.threadId ?? existing.messageId;
-  const recovered = await fetchAndUnarchiveThread(channel, threadId);
-  if (recovered) {
-    if (!existing.threadId) {
-      db.updateIssueThread(repo, issue.number, recovered.id);
-    }
-    return recovered;
-  }
-
-  const message = await withRetry(() => channel.messages.fetch(existing.messageId));
-  let thread: ThreadChannel;
-  try {
-    thread = await withRetry(() =>
-      message.startThread({
-        name: buildThreadName('Issue', issue.number, issue.title),
-        autoArchiveDuration: 1440,
-      })
-    );
-  } catch (error: unknown) {
-    if (isThreadAlreadyCreatedError(error)) {
-      const fallback = await fetchAndUnarchiveThread(channel, existing.messageId);
-      if (fallback) {
-        db.updateIssueThread(repo, issue.number, fallback.id);
-        return fallback;
-      }
-    }
-    throw error;
-  }
-
-  db.updateIssueThread(repo, issue.number, thread.id);
-
-  await withRetry(() => thread.send(`📋 Updates for Issue #${issue.number} will appear here.`));
-
-  return thread;
+  return getOrCreateMessageThread(
+    channel,
+    existing,
+    buildThreadName('Issue', issue.number, issue.title),
+    `📋 Updates for Issue #${issue.number} will appear here.`,
+    (threadId) => db.updateIssueThread(repo, issue.number, threadId)
+  );
 }
