+++ src/handlers/pr.ts
@@ -7,8 +7,12 @@ import { StateDb, PrMessage } from '../db/state.js';
 import { buildPrEmbed, buildPrComponents, buildMergedReply, buildClosedReply, buildPushReply, buildThreadName, PrData, ReviewStatus, CiStatus } from '../embeds/builders.js';
 import { getChannelForEvent, ChannelConfig } from '../config/channels.js';
 import { getExistingPrMessage } from '../discord/lookup.js';
+import { getOrCreateMessageThread } from '../discord/threads.js';
 import { withRetry } from '../utils/retry.js';
-import { isThreadAlreadyCreatedError, isUnknownMessageError } from '../utils/discord-errors.js';
+import { isUnknownMessageError } from '../utils/discord-errors.js';
+
+// Re-exported for compatibility (historically defined here)
+export { fetchAndUnarchiveThread } from '../discord/threads.js';
 
 export interface PrEventPayload {
   action: 'opened' | 'closed' | 'reopened' | 'synchronize' | 'edited' | 'ready_for_review' | 'converted_to_draft';
@@ -413,61 +417,11 @@ export async function getOrCreateThread(
   pr: PrData,
   existing: PrMessage
 ): Promise<ThreadChannel> {
-  // A message thread's ID equals its parent message's ID, so even when the
-  // DB has no threadId (channel-search recovery can't see archived threads —
-  // Message#thread is cache-only), the thread is still fetchable directly.
-  const threadId = existing.threadId ?? existing.messageId;
-  const recovered = await fetchAndUnarchiveThread(channel, threadId);
-  if (recovered) {
-    if (!existing.threadId) {
-      db.updatePrThread(repo, pr.number, recovered.id);
-    }
-    return recovered;
-  }
-
-  // Create a new thread on the message
-  const message = await withRetry(() => channel.messages.fetch(existing.messageId));
-  let thread: ThreadChannel;
-  try {
-    thread = await withRetry(() =>
-      message.startThread({
-        name: buildThreadName('PR', pr.number, pr.title),
-        autoArchiveDuration: 1440,
-      })
-    );
-  } catch (error: unknown) {
-    // 160004: the message already has a thread we couldn't see — fetch it
-    if (isThreadAlreadyCreatedError(error)) {
-      const fallback = await fetchAndUnarchiveThread(channel, existing.messageId);
-      if (fallback) {
-        db.updatePrThread(repo, pr.number, fallback.id);
-        return fallback;
-      }
-    }
-    throw error;
-  }
-
-  // Update the database with the new thread ID
-  db.updatePrThread(repo, pr.number, thread.id);
-
-  await withRetry(() => thread.send(`📋 Updates for PR #${pr.number} will appear here.`));
-
-  return thread;
-}
-
-/** Fetch a thread by ID and unarchive it; null if it doesn't exist. */
-export async function fetchAndUnarchiveThread(
-  channel: TextChannel,
-  threadId: string
-): Promise<ThreadChannel | null> {
-  try {
-    const thread = await withRetry(() => channel.threads.fetch(threadId));
-    if (!thread) return null;
-    if (thread.archived) {
-      await withRetry(async () => { await thread.setArchived(false); });
-    }
-    return thread;
-  } catch {
-    return null;
-  }
+  return getOrCreateMessageThread(
+    channel,
+    existing,
+    buildThreadName('PR', pr.number, pr.title),
+    `📋 Updates for PR #${pr.number} will appear here.`,
+    (threadId) => db.updatePrThread(repo, pr.number, threadId)
+  );
 }
